+++ okitweb/static/okit/js/okit_console.js
@@ -4,8 +4,10 @@
 */
 console.info('Loaded Console Javascript');
 
-const okitVersion = '0.27.1';
-const okitReleaseDate = '1st October 2021';
+
+const okitVersion = '0.28.0';
+const okitReleaseDate = '20th October 2021';
+
 // Validation
 const validate_error_colour = "#ff4d4d";
 const validate_warning_colour = "#ffd633";
@@ -115,7 +117,8 @@ function handleConfigChanged(event) {
     console_header_region_select.empty()
     console_header_region_select.append($('<option>').attr('value', '').text('Retrieving....'))
     okitRegions.load(okitSettings.profile)
-    okitOciData.load(okitSettings.profile)
+    okitOciData.load(okitSettings.profile, okitSettings.region)
+    setOCILink()
 }
 
 function loadHeaderRegionsDropDown() {
@@ -139,6 +142,7 @@ function handleRegionChanged(event) {
     event.stopPropagation()
     okitSettings.region = $('#console_header_region_select').val()
     okitSettings.save()
+    okitOciData.load(okitSettings.profile, okitSettings.region)
     setOCILink()
 }