+++ visualiser/query/ociQuery.py
@@ -35,6 +35,7 @@ class OCIQuery(OCIConnection):
         "AutonomousDatabase",
         "Backend",
         "BackendSet",
+        "Bastion",
         "BootVolume",
         "BootVolumeAttachment",
         "Bucket",
@@ -51,6 +52,7 @@ class OCIQuery(OCIConnection):
         "Export",
         "ExportSet",
         "FileSystem",
+        "Group",
         "Image",
         "Instance",
         # "InstancePool",
@@ -73,6 +75,8 @@ class OCIQuery(OCIConnection):
         "SecurityList",
         "ServiceGateway",
         "Subnet",
+        "User",
+        "UserGroupMembership",
         "Vcn",
         "VmCluster",
         "VmClusterNetwork",
@@ -83,6 +87,7 @@ class OCIQuery(OCIConnection):
     ]
     DISCOVERY_OKIT_MAP = {
         "AutonomousDatabase": "autonomous_databases",
+        "Bastion": "bastions",
         #"BootVolume": "block_storage_volumes",
         "Bucket": "object_storage_buckets",
         "Cluster": "oke_clusters",
@@ -95,6 +100,7 @@ class OCIQuery(OCIConnection):
         "Drg": "dynamic_routing_gateways",
         "ExadataInfrastructure": "exadata_infrastructures",
         "FileSystem": "file_storage_systems",
+        "Group": "groups",
         "Instance": "instances",
         "InstancePool": "instance_pools",
         "InternetGateway": "internet_gateways",
@@ -110,6 +116,7 @@ class OCIQuery(OCIConnection):
         "SecurityList": "security_lists",
         "ServiceGateway": "service_gateways",
         "Subnet": "subnets",
+        "User": "users",
         "Vcn": "virtual_cloud_networks",
         "VmCluster": "vm_clusters",
         "VmClusterNetwork": "vm_cluster_networks",
@@ -203,7 +210,10 @@ class OCIQuery(OCIConnection):
                         resource_list = self.route_tables(resource_list, resources)
                     elif resource_type == "ServiceGateway":
                         resource_list = self.service_gateways(resource_list, resources)
+                    elif resource_type == "Group":
+                        resource_list = self.groups(resource_list, resources)
                     # Check Life Cycle State
+                    # logger.info(f'Processing {resource_type} : {resource_list}')
                     response_json[self.DISCOVERY_OKIT_MAP[resource_type]] = [r for r in resource_list if "lifecycle_state" not in r or r["lifecycle_state"] in self.VALID_LIFECYCLE_STATES]
                     #response_json[self.DISCOVERY_OKIT_MAP[resource_type]] = resource_list
         return response_json
@@ -225,6 +235,11 @@ class OCIQuery(OCIConnection):
                 ess = [e for e in export_sets if e["id"] == mt["export_set_id"]]
                 mt["export_set"] = ess[0] if len(ess) else {}
         return file_storage_systems
+
+    def groups(self, groups, resources):
+        for group in groups:
+            group["user_ids"] = [m["user_id"] for m in resources.get("UserGroupMembership", []) if m["group_id"] == group["id"]]
+        return groups
 
     def instances(self, instances, resources):
         # Exclude OKE Instances